+++ src/FairPlayCombinedSln/FairPlayCombined.DataAccess/Models/FairPlayTube/VideoInfo.cs
@@ -56,7 +56,7 @@ public partial class VideoInfo
     [StringLength(450)]
     public string ApplicationUserId { get; set; }
 
-    public short VideoIndexStatusId { get; set; }
+    public int VideoIndexStatusId { get; set; }
 
     public double VideoDurationInSeconds { get; set; }
