+++ src/FairPlayCombinedSln/FairPlayCombined.Models/FairPlayTube/VideoInfo/VideoInfoModel.cs
@@ -21,13 +21,13 @@ namespace FairPlayCombined.Models.FairPlayTube.VideoInfo
         public string? VideoBloblUrl { get; set; }
         public string? IndexedVideoUrl { get; set; }
         public string? ApplicationUserId { get; set; }
-        public short VideoIndexStatusId { get; set; }
+        public int VideoIndexStatusId { get; set; }
         public double VideoDurationInSeconds { get; set; }
         public string? VideoIndexSourceClass { get; set; }
         public decimal Price { get; set; }
         public string? ExternalVideoSourceUrl { get; set; }
         public string? VideoLanguageCode { get; set; }
-        public short VideoVisibilityId { get; set; }
+        public int VideoVisibilityId { get; set; }
         public string? ThumbnailUrl { get; set; }
         public string? YouTubeVideoId { get; set; }
     }
