+++ src/FairPlayCombinedSln/FairPlayCombinedDb/FairPlayTube/Tables/VideoIndexStatus.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [FairPlayTube].[VideoIndexStatus]
 (
-	[VideoIndexStatusId] SMALLINT NOT NULL CONSTRAINT PK_VideoIndexStatus PRIMARY KEY, 
+	[VideoIndexStatusId] INT NOT NULL CONSTRAINT PK_VideoIndexStatus PRIMARY KEY, 
     [Name] NVARCHAR(50) NOT NULL
 )
